+++ test/src/main/java/alehad/cloud/test/msg/model/Message.java
@@ -2,22 +2,22 @@ package alehad.cloud.test.msg.model;
 
 public class Message {
 
-	private long 	id;
+	private int 	id;
 	private String 	message;
 	
 	public Message() {
 		//no-op default constructor
 	}
 	
-	public Message(long id, String message) {
+	public Message(int id, String message) {
 		this.id 	 = id;
 		this.message = message;
 	}
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
